users/views.py
from rest_framework import generics
from users.models import User
from users.serializers import UserSerializer

class UserListAPIView(generics.ListAPIView):
    serializer_class = UserSerializer
    queryset = User.objects.all()